+++ descope/management/user_async.py
@@ -1586,6 +1586,25 @@ class UserAsync(UserBase, AsyncHTTPBase):
         )
         return
 
+    async def remove_recovery_codes(
+        self,
+        login_id: str,
+    ) -> None:
+        """
+            Removes all recovery codes for the user with the given login ID or user ID.
+
+        Args:
+        login_id (str): The login ID or user ID of the user to remove recovery codes for.
+
+        Raise:
+        AuthException: raised if the operation fails
+        """
+        await self._http.post(
+            MgmtV1.user_remove_recovery_codes_path,
+            body={"loginId": login_id},
+        )
+        return
+
     async def generate_otp_for_test_user(
         self,
         method: DeliveryMethod,